requirements.txt
flask
flask-cors
pandas
matplotlib
numpy
python-dotenv
groq
gunicorn
